优化group查找
**复现步骤**

* 无

**原因和解决方案**

* acl中的groups再分组，减少生成map时用的时间

**影响范围**

* 无

**验证方案**

* 验证通过

**涉及分支**

* master

**检查项**

- [ ] 需要更新dep依赖
- [ ] 是共性问题(代码中存在类似问题)
- [x] 编译通过
- [x] 单元测试通过

/assign @buyi

/cc @buyi

+++ datatype/endpoint.go
@@ -76,7 +76,7 @@ type LookupKey struct {
 	Tap                      TapType
 	Invalid                  bool
 	FastIndex                int
-	SrcGroupIds, DstGroupIds []uint32
+	SrcGroupIds, DstGroupIds []uint16
 	FeatureFlag              FeatureFlags
 }
